% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandCategorizationWeight.R
\name{calcLandCategorizationWeight}
\alias{calcLandCategorizationWeight}
\title{calcLandCategorizationWeight}
\usage{
calcLandCategorizationWeight(map, geometry, crs)
}
\arguments{
\item{map}{a map in form of a 2 column data.frame containing a mapping between
input categories and output categories for the data dimension of a magpie object}

\item{geometry}{the geometry of the magpie object for which the categories should
be mapped as given in the geometry attribute of a magpie object with
geometry information \code{attr(x, "geometry")}.}

\item{crs}{the coordinate reference system as returned by \code{attr(x, "crs")} from
a magpie object with coordinates information.}
}
\description{
Computes weights for a weighted category mapping and aggregates the weights
to the spatial resolution described by the arguments \code{geometry} and \code{crs}.
}
\note{
This calc-function has a rather unusual shape in that sense that the arguments to
be provided are not simple configuration settings but rather relative complex. For this
kind of implementation it is typically advised to use tool functions. The reason that a
calc function is used in this particular case instead is, that the inputs of this
function usually do not change over a long period of time so that caching of results
becomes key for the overall performance of the data processing, which is available
for calc- but not tool-functions. If tool-functions might support caching in the future
as well as a conversion to a tool function might become a sensible option.
}
\author{
Jan Philipp Dietrich
}
