% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandHighRes.R
\name{calcLandHighRes}
\alias{calcLandHighRes}
\title{calcLandHighRes}
\usage{
calcLandHighRes(
  input = "magpie",
  target = "luh2mod",
  harmonizationPeriod = c(2015, 2050),
  downscaling = "magpieClassic"
)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, currently only "luh2"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{downscaling}{name of downscaling method, currently only "magpieClassic"}
}
\value{
downscaled land use data
}
\description{
This function performs the downscaling: It calculates a high resolution dataset
from the low resolution input dataset and the high resolution target dataset
using the given downscaling method.
}
\author{
Jan Philipp Dietrich
}
