% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandReport.R
\name{calcLandReport}
\alias{calcLandReport}
\title{calcLandReport}
\usage{
calcLandReport(project = "RESCUE")
}
\arguments{
\item{project}{name of the project, currently only "RESCUE"}
}
\value{
land use data
}
\description{
Convert the downscaled land use data to the format required by the given project.
}
\author{
Pascal Sauer
}
