% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTarget.R
\name{calcLandTarget}
\alias{calcLandTarget}
\title{calcLandTarget}
\usage{
calcLandTarget(target = "luh2mod")
}
\arguments{
\item{target}{name of the target dataset, options are: luh2, luh2mod}
}
\value{
land target data
}
\description{
Prepare the high resolution target land use dataset for
harmonization and downscaling, checking data for consistency before returning.
}
\author{
Pascal Sauer
}
