% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTargetData.R
\name{calcLandTargetData}
\alias{calcLandTargetData}
\title{calcLandTargetData}
\usage{
calcLandTargetData(target = "luh2")
}
\arguments{
\item{target}{name of the target dataset, currently only "luh2"}
}
\value{
land target data
}
\description{
Prepare the high resolution target land use dataset for
harmonization and downscaling, checking data for consistency before returning.
}
\author{
Pascal Führlich
}
