% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTargetExtrapolated.R
\name{calcLandTargetExtrapolated}
\alias{calcLandTargetExtrapolated}
\title{calcLandTargetExtrapolated}
\usage{
calcLandTargetExtrapolated(
  input = "magpie",
  target = "luh2mod",
  transitionYears = seq(2020, 2045, 5)
)
}
\arguments{
\item{input}{character, name of the input data set, currently only "magpie"}

\item{target}{character, name of the target data set, currently only "luh2mod"}

\item{transitionYears}{years to which the target data is extrapolated}
}
\value{
extrapolated land target data, if calcOutput is called with
supplementary = TRUE wood harvest area is also returned
}
\description{
Aggregated low resolution target data is extrapolated to the given years
using toolExtrapolate and normalized afterwards, so that the total sum over
all land types is unchanged.
To account for the relationship between wood
harvest area and primary land (which is no longer primary once it has been
harvested) wood harvest area is calculated here even though it is otherwise
considered a nonland variable. The share of primary land that was
harvested in the historical period is calculated and then multiplied
by the maximum possible harvest in the extrapolation period. Primary land
is then converted to secondary land so the total reduction equals the area
that was harvested.
}
\author{
Pascal Sauer
}
