% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTargetLowRes.R
\name{calcLandTargetLowRes}
\alias{calcLandTargetLowRes}
\title{calcLandTargetLowRes}
\usage{
calcLandTargetLowRes(input, target)
}
\arguments{
\item{input}{name of an input dataset, see \code{\link{calcLandInput}}
available input datasets}

\item{target}{name of a target dataset, see \code{\link{calcLandTarget}}
available target datasets}
}
\value{
low resolution target land data
}
\description{
Aggregate target land data to the spatial resolution of the input data in
preparation for harmonization.
}
\author{
Pascal Sauer
}
