% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTransitions.R
\name{calcLandTransitions}
\alias{calcLandTransitions}
\title{calcLandTransitions}
\usage{
calcLandTransitions(
  outputFormat = "ESM",
  harmonizationPeriod = c(2015, 2050),
  gross = TRUE
)
}
\arguments{
\item{outputFormat}{format in which the outputs should be prepared. Currently,
only "ESM" for earth system model compatible input data is available.}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{gross}{either boolean or a magpie object containing bidirectional
transition shares relative to the area of the involved land pools (transition
divided by the area of the land pool in the "from" sub dimension). If set to
FALSE only net transitions will be returned. If set to TRUE an internal
gross transition estimate based on average gross transitions in LUH2 in the
period from 1995 to 2015 will be used.}
}
\value{
land use transition data
}
\description{
Ex-Post estimation of land use transitions based on land use state
information
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
