% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHarmonized.R
\name{calcNonlandHarmonized}
\alias{calcNonlandHarmonized}
\title{calcNonlandHarmonized}
\usage{
calcNonlandHarmonized(
  input = "magpie",
  target = "luh2mod",
  harmonizationPeriod = c(2015, 2050),
  method = "fade"
)
}
\arguments{
\item{input}{name of the input dataset, currently only "magpie"}

\item{target}{name of the target dataset, currently only "luh2"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{method}{harmonization method, see \code{\link{toolGetHarmonizer}} for available methods}
}
\value{
harmonized nonland data
}
\description{
Harmonize nonland input data to target data using the specified method, checking
data for consistency before returning.
}
\details{
Wood harvest biomass (bioh) is adapted to the harmonized wood harvest area
by calculating kg C per mega hectare for input and target data and
harmonizing it. This is then multiplied by the harmonized wood harvest area
and scaled so the total harmonized bioh is reached.
}
\author{
Pascal Sauer
}
