% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHarmonizedCategories.R
\name{calcNonlandHarmonizedCategories}
\alias{calcNonlandHarmonizedCategories}
\title{calcNonlandHarmonizedCategories}
\usage{
calcNonlandHarmonizedCategories(
  input = "magpie",
  target = "luh2mod",
  youngShareWoodHarvestArea = 0.95,
  youngShareWoodHarvestWeight = 0.5
)
}
\arguments{
\item{input}{name of the input dataset, currently only "magpie"}

\item{target}{name of the target dataset, currently only "luh2"}

\item{youngShareWoodHarvestArea}{share of wood harvest area taken from young (instead of mature) secondary forest;
default value is based on LUH value from 2014; used to disaggregate wood harvest area from secondary forest to
secondary young and mature forest}

\item{youngShareWoodHarvestWeight}{analogue to youngShareWoodHarvestArea for wood harvest weight instead of area}
}
\value{
nonland data with target categories
}
\description{
Harmonize categories by mapping nonland input data categories to the categories of the nonland target dataset.
See \code{\link{calcLandHarmonizedCategories}} for an explanation of the mapping procedure.
}
\author{
Pascal Sauer
}
