% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHighRes.R
\name{calcNonlandHighRes}
\alias{calcNonlandHighRes}
\title{calcNonlandHighRes}
\usage{
calcNonlandHighRes(
  input = "magpie",
  target = "luh2mod",
  harmonizationPeriod = c(2015, 2050)
)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, currently only "luh2mod"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
downscaled nonland data
}
\description{
Calculate a high resolution dataset from the low resolution input dataset
and high resolution data.
}
\details{
Wood harvest area is disaggregated using the maximum possible harvest per
year which is based on the downscaled land data. Bioh
is disaggregated using the just disaggregated wood harvest area as weight.
Fertilizer is disaggregated using the downscaled land data. Harvest weight
type is disaggregated using the nonland target data in the first year of
the harmonization period as weight.
}
\author{
Pascal Sauer
}
