% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandInputData.R
\name{calcNonlandInputData}
\alias{calcNonlandInputData}
\title{calcNonlandInputData}
\usage{
calcNonlandInputData(input = "magpie")
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}
}
\value{
nonland input data
}
\description{
Prepare the nonland input data for category mapping, checking data for consistency before returning.
}
\details{
All "Land" functions deal with area data, as opposed to "Nonland" functions which deal with non-area
data such as the amount of applied fertilizer. These are treated differently, because for area
data other constraints apply, e.g. the total area must be constant over time.
}
\author{
Pascal Sauer
}
