% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandReport.R
\name{calcNonlandReport}
\alias{calcNonlandReport}
\title{calcNonlandReport}
\usage{
calcNonlandReport(project = "RESCUE")
}
\arguments{
\item{project}{name of the project, currently only "RESCUE"}
}
\value{
nonland data
}
\description{
Convert the downscaled nonland data to the format required by the given project.
}
\author{
Pascal Sauer
}
