% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandTarget.R
\name{calcNonlandTarget}
\alias{calcNonlandTarget}
\title{calcNonlandTarget}
\usage{
calcNonlandTarget(target = "luh2mod")
}
\arguments{
\item{target}{name of the target dataset, currently only "luh2"}
}
\value{
nonland target data
}
\description{
Prepare the high resolution nonland target dataset for
harmonization and downscaling, checking data for consistency before returning.
}
\author{
Pascal Sauer
}
