% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandTargetExtrapolated.R
\name{calcNonlandTargetExtrapolated}
\alias{calcNonlandTargetExtrapolated}
\title{calcNonlandTargetExtrapolated}
\usage{
calcNonlandTargetExtrapolated(
  input = "magpie",
  target = "luh2mod",
  transitionYears = seq(2020, 2045, 5)
)
}
\arguments{
\item{input}{character, name of the input data set, currently only "magpie"}

\item{target}{character, name of the target data set, currently only "luh2mod"}

\item{transitionYears}{years to which the target data is extrapolated}
}
\value{
extrapolated nonland target data
}
\description{
Aggregated low resolution target data is extrapolated to the given years
using toolExtrapolate. To extrapolate wood harvest weight (bioh) multiply
wood harvest area already extrapolated by calcLandTargetExtrapolated with
the historical wood harvest rate in kg C per Mha.
}
\author{
Pascal Sauer
}
