% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResolutionMapping.R
\name{calcResolutionMapping}
\alias{calcResolutionMapping}
\title{calcResolutionMapping}
\usage{
calcResolutionMapping(input = "magpie", target = "luh2mod")
}
\arguments{
\item{input}{character, the input dataset, currently only "magpie" is supported}

\item{target}{character, the target dataset, currently only "luh2mod" is supported}
}
\value{
a list including a data.frame with columns x, y, lowRes, countrycode
}
\description{
Calculate a complete mapping from low (input dataset, clusters/countries/regions)
to high resolution (target dataset, grid). As a basis the mapping from the low
resolution clusters/countries/regions to grid cells is used.
Cells which are present in that mapping, but not in the target
dataset are discarded. Cells which are present in the target dataset, but not in the mapping are
added using a nearest-neighbor approach: These cells are mapped to the same low resolution
cluster/country/region as the closest cell which is already present in the mapping.
}
\author{
Pascal Sauer
}
