% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTransitionsHighRes.R
\name{calcTransitionsHighRes}
\alias{calcTransitionsHighRes}
\title{calcTransitionHighRes}
\usage{
calcTransitionsHighRes(
  input = "magpie",
  target = "luh2",
  diffYears = c(1995, 2000, 2005, 2010)
)
}
\arguments{
\item{input}{name of the land input source to be used (default "magpie")}

\item{target}{name of the land target source to be used (default "luh2")}

\item{diffYears}{year between the transitions should be calculated. As these
computations can be quite memory consumed it is advised not to select too many
years at once}
}
\description{
Computes a land use transition data set for a given input as as well as
target data source.
}
\author{
Jan Philipp Dietrich
}
