% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLUH2v2h.R
\name{convertLUH2v2h}
\alias{convertLUH2v2h}
\title{convertLUH2v2h}
\usage{
convertLUH2v2h(x, subtype = "states")
}
\arguments{
\item{x}{SpatRaster with LUH2 cell area shares}

\item{subtype}{Only "states" and "woodHarvestArea" are converted}
}
\description{
Convert LUH2 cell area shares to absolute areas by multiplying with cell area, convert to Mha.
}
