% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLUH2v2h.R
\name{downloadLUH2v2h}
\alias{downloadLUH2v2h}
\title{downloadLUH2v2h}
\usage{
downloadLUH2v2h(subtype = NULL)
}
\arguments{
\item{subtype}{one of states, management, transitions, cellArea}
}
\value{
metadata list with URL, DOI, title, description, author, unit, version, release date
}
\description{
Download the LUH2v2h dataset (states.nc, transitions.nc,
management.nc, staticData_quarterdeg.nc).
}
\author{
Pascal Sauer
}
