% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleRun.R
\name{downscaleRun}
\alias{downscaleRun}
\title{downscaleRun}
\usage{
downscaleRun(outputfolder, ...)
}
\arguments{
\item{outputfolder}{path to a folder containing fulldata.gdx and clustermap_*.rds,
resulting tgz will be written here}

\item{...}{arguments passed on to retrieveData,
e.g. model ("DOWNSCALEDMAGPIE" or "ESM"), rev, harmonizationPeriod}
}
\value{
Invisibly, the path to the newly created tgz archive.
}
\description{
Downscale MAgPIE results from the given folder. The resulting tgz file will
be generated in the given folder.
}
\author{
Pascal Sauer
}
