% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleRunESM.R
\name{downscaleRunESM}
\alias{downscaleRunESM}
\title{downscaleRunESM}
\usage{
downscaleRunESM(outputdir, revision = NULL, scenario = NULL, ...)
}
\arguments{
\item{outputdir}{path to a folder containing fulldata.gdx and clustermap_*.rds,
resulting tgz will be written here}

\item{revision}{passed on to retrieveData, default: current date}

\item{scenario}{passed on to retrieveData, default: slightly modified folder name}

\item{...}{additional arguments passed on to retrieveData}
}
\value{
Invisibly, the path to the newly created tgz archive.
}
\description{
Downscale MAgPIE results from the given folder to 0.25 degree resolution in
LUH2 format for ESMs. The resulting tgz file will be generated in the given folder.
}
\author{
Pascal Sauer
}
