% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullDOWNSCALEDMAGPIE.R
\name{fullDOWNSCALEDMAGPIE}
\alias{fullDOWNSCALEDMAGPIE}
\title{fullDOWNSCALEDMAGPIE}
\usage{
fullDOWNSCALEDMAGPIE(
  rev = numeric_version("0"),
  ...,
  harmonizationPeriod = c(2015, 2050)
)
}
\arguments{
\item{rev}{revision number of the data. If not provided the current date will be used instead.
When called via madrat::retrieveData rev will be converted to numeric_version.}

\item{...}{reserved for future use}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\description{
Run the pipeline to generate harmonized and downscaled MAgPIE data using
landuseinit as a reference dataset. Write output in the format of
avl_land_t_0.5.mz, full report on consistency checks is printed and
written to report.log.
}
\author{
Pascal Sauer
}
