% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullESM.R
\name{fullESM}
\alias{fullESM}
\title{fullESM}
\usage{
fullESM(
  rev = NULL,
  ...,
  scenario = "",
  years = c(2020, 2100),
  harmonizationPeriod = c(2015, 2050),
  compression = 2,
  progress = TRUE
)
}
\arguments{
\item{rev}{revision number of the data. If not provided the current date will be used instead.
When called via madrat::retrieveData rev will be converted to numeric_version.}

\item{...}{reserved for future use}

\item{scenario}{scenario name to be included in filenames}

\item{years}{vector of min and max year specifying the range of years for which
years will be returned. Years outside of this range will be ignored.}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}

\item{progress}{boolean defining whether progress should be printed}
}
\description{
Run the pipeline to generate harmonized and downscaled data to report for the RESCUE, OptimESM
and other projects where ESM compatible land use inputs are required.
Write .nc files, print full report on consistency checks and write it to report.log.
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
