% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRESCUE.R
\name{fullRESCUE}
\alias{fullRESCUE}
\title{fullRESCUE}
\usage{
fullRESCUE()
}
\description{
Run the pipeline to generate harmonized and downscaled data to report for the RESCUE project.
Data is calculated for every 5th year, fill years inbetween with linear interpolation.
Write .nc files, print full report on consistency checks and write it to report.log.
}
\author{
Pascal Führlich
}
