% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmagpiedownscaled-package.R
\docType{package}
\name{mrmagpiedownscaled-package}
\alias{mrmagpiedownscaled}
\alias{mrmagpiedownscaled-package}
\title{mrmagpiedownscaled: Downscale and harmonize MAgPIE output using LUH2 as reference data}
\description{
Downscale and harmonize MAgPIE output using LUH2 as reference data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrmagpiedownscaled}
}

}
\author{
\strong{Maintainer}: Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}

}
