% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLowHigh.R
\name{plotLowHigh}
\alias{plotLowHigh}
\title{plotLowHigh}
\usage{
plotLowHigh(variable = "c3ann", year = 2040, range = c(0, 1))
}
\arguments{
\item{variable}{character, a regex that should match at least one variable
name in LandHighRes and LandHarmonized, if multiple are matched they are
summed up}

\item{year}{integer, year to plot}

\item{range}{range for plotting}
}
\description{
Plot a land variable in low (after harmonization) and high resolution (after
downscaling) for a given year, unit is area share.
}
\author{
Pascal Sauer
}
