% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMagpieFulldataGdx.R
\name{readMagpieFulldataGdx}
\alias{readMagpieFulldataGdx}
\title{readMagpieFulldataGdx}
\usage{
readMagpieFulldataGdx(subtype = "land", subset = 1995:2100)
}
\arguments{
\item{subtype}{type of data to be read in. Available options are land,
crop, woodHarvestWeight, woodHarvestArea and fertilizer}

\item{subset}{Available years (usually timestep is 5+ years) are only returned if they are in subset.}
}
\description{
Read function for data coming from the MAgPIE model.
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
