% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAddMetadataESM.R
\name{toolAddMetadataESM}
\alias{toolAddMetadataESM}
\title{toolAddMetadataESM}
\usage{
toolAddMetadataESM(ncFile, cfg)
}
\arguments{
\item{ncFile}{file name of the respective nc file}

\item{cfg}{list of settings containing the elements "revision", "missingValue",
"resolution", "compression" and "harmonizationPeriod"}
}
\description{
Add metdata to ESM compatible nc output files
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
