% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCheckWoodHarvestArea.R
\name{toolCheckWoodHarvestArea}
\alias{toolCheckWoodHarvestArea}
\title{toolCheckWoodHarvestArea}
\usage{
toolCheckWoodHarvestArea(harvest, land, notePrefix = "")
}
\arguments{
\item{harvest}{magpie object with exactly the following categories:
paste0(c("primf", "secyf", "secmf", "primn", "secnf"), "_wood_harvest_area")}

\item{land}{magpie object with at least the following categories:
c("primf", "secdf", "forestry", "primn", "secdn")}

\item{notePrefix}{character to prepend to the check's message}
}
\description{
Check wood harvest area is not exceeding land area of the corresponding
type divided by timestep length. Also, check that primf and primn
are reduced by at least as much as they were harvested.
}
\author{
Pascal Sauer
}
