% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDownscaleMagpieClassic.R
\name{toolDownscaleMagpieClassic}
\alias{toolDownscaleMagpieClassic}
\title{toolDownscaleMagpieClassic}
\usage{
toolDownscaleMagpieClassic(x, xTarget, mapping)
}
\arguments{
\item{x}{magclass containing land to be downscaled}

\item{xTarget}{magclass target land use dataset for initialization year}

\item{mapping}{mapping between \code{x} and \code{xTarget}}
}
\value{
downscaled land use dataset
}
\description{
classic MAgPIE downscaling method using \code{luscale::interpolate2} as
downscaling function.
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
