% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFillYearsSpatRaster.R
\name{toolFillYearsSpatRaster}
\alias{toolFillYearsSpatRaster}
\title{toolFillYearsSpatRaster}
\usage{
toolFillYearsSpatRaster(x, years = NULL)
}
\arguments{
\item{x}{SpatRaster with years in layer names}

\item{years}{data for these years will be added if they are not already present,
if years is NULL all years between the first and last year in x will be filled}
}
\value{
SpatRaster with filled years
}
\description{
Fill data for missing years using linear interpolation.
}
\author{
Pascal Führlich
}
