% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolGetSmallerArea.R
\name{toolGetSmallerArea}
\alias{toolGetSmallerArea}
\title{toolGetSmallerArea}
\usage{
toolGetSmallerArea(states)
}
\arguments{
\item{states}{magpie dataset containing states information}
}
\value{
A land type x land type data set containing for each possible combination
the smaller area of the two land types
}
\description{
The function creates a transition data set based on a state dataset which
contains for every possible connection (land type A <-> land type B) the
area of the smaller of these land types.
This information is relevant if transitional shares should be calculated
based on the smaller area of the two (assuming that the smaller area
is determining how much transistions there will be).
}
\author{
Jan Philipp Dietrich
}
