% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeOffset.R
\name{toolHarmonizeOffset}
\alias{toolHarmonizeOffset}
\title{toolHarmonizeOffset}
\usage{
toolHarmonizeOffset(input, target, harmonizationPeriod)
}
\arguments{
\item{input}{magpie object to fade to, usually model projections}

\item{target}{magpie object to fade from, usually historical data, this argument is called "target"
for consistency with argument names of other functions, it is somewhat misleading here,
because data after the harmonizationPeriod will not be taken from target, but from input}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
magpie object with harmonized data
}
\description{
Harmonize datasets using the offset method as implemented in mip::harmonize,
originally implemented in Python as part of aneris.
}
\author{
Pascal Sauer
}
