% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLandCategoriesMapping.R
\name{toolLandCategoriesMapping}
\alias{toolLandCategoriesMapping}
\title{toolLandCategoriesMapping}
\usage{
toolLandCategoriesMapping(input, target)
}
\arguments{
\item{input}{name of the land input source to be used (default "magpie")}

\item{target}{name of the land target source to be used (default "luh2")}
}
\description{
Computes a land category mapping between input and target data set via
retrieving mappings from these sources to an internally defined
reference categorization and merging them to a direct input-to-target
mapping.
Mappings to reference categories are stored internally in the package and
have to be added to it if new input and/or target data should be supported.
}
\author{
Jan Philipp Dietrich
}
