% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolMaxExpansion.R
\name{toolMaxExpansion}
\alias{toolMaxExpansion}
\title{toolMaxExpansion}
\usage{
toolMaxExpansion(x, variable, removeNA = FALSE)
}
\arguments{
\item{x}{A magclass object}

\item{variable}{A character string indicating the variable to be used}

\item{removeNA}{A logical indicating whether NA values should be ignored}
}
\value{
A numeric value indicating the maximum expansion of the variable
}
\description{
Return the maximum expansion of a variable from one timestep to the next.
}
\author{
Pascal Sauer
}
