% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrimExpansionCheck.R
\name{toolPrimExpansionCheck}
\alias{toolPrimExpansionCheck}
\title{toolPrimExpansionCheck}
\usage{
toolPrimExpansionCheck(x)
}
\arguments{
\item{x}{A magclass object}
}
\description{
Check that primary forest (primf) and primary nature/nonland (primf) are
never expanding using toolExpectTrue. If they are expanding, a warning
including the maximum expansion is thrown. Categories not present
in x are skipped.
}
\author{
Pascal Sauer
}
