% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolReplaceExpansion.R
\name{toolReplaceExpansion}
\alias{toolReplaceExpansion}
\title{toolReplaceExpansion}
\usage{
toolReplaceExpansion(
  x,
  from,
  to,
  ...,
  noteThreshold = 10^-10,
  warnThreshold = 10^-5
)
}
\arguments{
\item{x}{a magpie object}

\item{from}{name of a land category, e.g. "primf"
Expansion of this category happening from one timestep to the next
will be replaced. If missing in x return x unchanged}

\item{to}{name of another land category, e.g. "secdf"
Expansion of 'from' will be replaced with expansion of 'to'}

\item{...}{not used, will throw an error if supplied}

\item{noteThreshold}{expansion greater than this will trigger a note
that expansion was replaced}

\item{warnThreshold}{expansion greater than this will trigger a warning
that expanding considerably}
}
\value{
a magpie object with expansion of 'from' replaced by 'to'
}
\description{
Expansion from one timestep to the next of one land type is replaced with
another land type.
}
\author{
Pascal Sauer
}
