% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolScaleConstantArea.R
\name{toolScaleConstantArea}
\alias{toolScaleConstantArea}
\title{toolScaleConstantArea}
\usage{
toolScaleConstantArea(x, ..., noteThreshold = 10^-10, warnThreshold = 10^-5)
}
\arguments{
\item{x}{a magpie object}

\item{...}{not used, will throw an error if supplied}

\item{noteThreshold}{if the maximum difference between the scaled and the
original data is greater than this, a note is triggered}

\item{warnThreshold}{if the maximum difference between the scaled and the
original data is greater than this, a warning is triggered}
}
\value{
x scaled
}
\description{
Scale x to make the sum of all land types is constant over time.
}
\author{
Pascal Sauer
}
