% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSpatRasterToDataset.R
\name{toolSpatRasterToDataset}
\alias{toolSpatRasterToDataset}
\title{toolSpatRasterToDataset}
\usage{
toolSpatRasterToDataset(x)
}
\arguments{
\item{x}{SpatRaster with names of the form "y[0-9]+..[varname]"}
}
\value{
SpatRasterDataset
}
\description{
Convert a SpatRaster to a SpatRasterDataset.
}
\author{
Pascal Führlich
}
