% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolTransitionsBasic.R
\name{toolTransitionsBasic}
\alias{toolTransitionsBasic}
\title{toolTransitionsBasic}
\usage{
toolTransitionsBasic(x)
}
\arguments{
\item{x}{magpie dataset containing land data with at least two time steps
to extract net transitions from}
}
\description{
tool function to extract net transitions between categories from a land
dataset with at least 2 time steps. The approach is rather simplistic by
assuming that expansion happens proportionally in all affected classes (equal
preference of transistions across all categories).
}
\author{
Jan Philipp Dietrich
}
