% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWarnIfExpansion.R
\name{toolWarnIfExpansion}
\alias{toolWarnIfExpansion}
\title{toolWarnIfExpansion}
\usage{
toolWarnIfExpansion(x, variable)
}
\arguments{
\item{x}{A magclass object}

\item{variable}{A vector of variables to be checked}
}
\description{
Warn if expansion is detected one of the given variables, giving the
maximum expansion in each respective variable.
}
\author{
Pascal Sauer
}
