% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWoodHarvestArea.R
\name{toolWoodHarvestArea}
\alias{toolWoodHarvestArea}
\title{toolWoodHarvestArea}
\usage{
toolWoodHarvestArea(x, land, fix)
}
\arguments{
\item{x}{magpie object with exactly the following categories:
paste0(c("primf", "secyf", "secmf", "primn", "secnf"), "_wood_harvest_area")}

\item{land}{magpie object with at least the following categories:
c("primf", "secdf", "primn", "secdn")}

\item{fix}{logical, if TRUE reduce wood harvest area to at most match land area}
}
\value{
potentially fixed magpie object with the same dimensions as x
}
\description{
Check wood harvest area is not exceeding land area of the corresponding
type in the previous timestep divided by timestep length. Optionally fix
by simply reducing wood harvest area to at most match land area.
}
\author{
Pascal Sauer
}
