% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWriteNC.R
\name{toolWriteNC}
\alias{toolWriteNC}
\title{toolWriteNC}
\usage{
toolWriteNC(x, variables, fileName, now = Sys.time(), compression = 2)
}
\arguments{
\item{x}{data to be written to file}

\item{variables}{layers that should be selected from \code{x}}

\item{fileName}{file name that should be used for the resulting nc file}

\item{now}{time that should be used as time stamp in metadata}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}
}
\description{
Tool function to write LUH2-style NetCDF files
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
