% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWriteStates.R
\name{toolWriteStates}
\alias{toolWriteStates}
\title{toolWriteManagement}
\usage{
toolWriteStates(land, fileSuffix, now = Sys.time(), compression = 2)
}
\arguments{
\item{land}{land data the management information should be extracted from}

\item{fileSuffix}{second half of the resulting file name for the output NetCDF file}

\item{now}{time that should be used as time stamp in metadata}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}
}
\description{
Tool function to write LUH2-style states data in NetCDF format
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
