% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWriteTransitions.R
\name{toolWriteTransitions}
\alias{toolWriteTransitions}
\title{toolWriteTransitions}
\usage{
toolWriteTransitions(
  transitions,
  nonland,
  fileSuffix,
  now = Sys.time(),
  compression = 2,
  interpolate = FALSE
)
}
\arguments{
\item{transitions}{land transition data the information should be extracted from}

\item{fileSuffix}{second half of the resulting file name for the output NetCDF file}

\item{now}{time that should be used as time stamp in metadata}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}

\item{interpolate}{boolean defining whether the data should be interpolated to annual values or not}
}
\description{
Tool function to write LUH2-style land transition data in NetCDF format
}
\author{
Jan Philipp Dietrich
}
