% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcESMStates.R
\name{calcESMStates}
\alias{calcESMStates}
\title{calcESMStates}
\usage{
calcESMStates(harmonizationPeriod, yearsSubset)
}
\arguments{
\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{remove years from the returned data which are not in yearsSubset}
}
\value{
ESM compatible states information
}
\description{
Create ESM compatible state information
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
