% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcESMTransitions.R
\name{calcESMTransitions}
\alias{calcESMTransitions}
\title{calcESMTransitions}
\usage{
calcESMTransitions(harmonizationPeriod, yearsSubset)
}
\arguments{
\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsSubset}{remove years from the returned data which are not in yearsSubset}
}
\value{
ESM compatible transistions information
}
\description{
Create ESM compatible transitions information
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
