% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandHarmonized.R
\name{calcLandHarmonized}
\alias{calcLandHarmonized}
\title{calcLandHarmonized}
\usage{
calcLandHarmonized(
  input = "magpie",
  target = "luh2",
  harmonizeYear = 2015,
  finalYear = 2050,
  method = "extrapolateFade"
)
}
\arguments{
\item{input}{name of the land input source to be used (default "magpie")}

\item{target}{name of the land target source to be used (default "luh2")}

\item{harmonizeYear}{year in which the transition from target to input
data begins}

\item{finalYear}{year in which the transition shall be completed}

\item{method}{transitioning method (default "extrapolateFade")}
}
\description{
This function computes a version of the chosen land input data in the resolution
of the land input data but harmonized to the land use information of the
chosen land target data set (harmonized categories as well as harmonized
transition from historic target data to simulated input data).
}
\author{
Pascal Führlich, Jan Philipp Dietrich
}
