% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandHighRes.R
\name{calcLandHighRes}
\alias{calcLandHighRes}
\title{calcLandHighRes}
\usage{
calcLandHighRes(
  input,
  target,
  harmonizationPeriod,
  yearsToKeep,
  downscaling = "magpieClassic"
)
}
\arguments{
\item{input}{name of an input dataset}

\item{target}{name of a target dataset}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}

\item{yearsToKeep}{vector of years to keep in the output dataset}

\item{downscaling}{name of downscaling method, currently only "magpieClassic"}
}
\value{
downscaled land use data
}
\description{
This function performs the downscaling: It calculates a high resolution dataset
from the low resolution input dataset and the high resolution target dataset
using the given downscaling method.
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
