% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandInput.R
\name{calcLandInput}
\alias{calcLandInput}
\title{calcLandInput}
\usage{
calcLandInput(input = "magpie")
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}
}
\value{
land input data
}
\description{
Prepare the land input data for the category mapping, checking data for consistency before returning.
All "Land" functions deal with area data, as opposed to "Nonland" functions which deal with non-area
data such as the amount of applied fertilizer. These are treated differently, because for area
data other constraints apply, e.g. the total area must be constant over time.
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
