% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandReport.R
\name{calcLandReport}
\alias{calcLandReport}
\title{calcLandReport}
\usage{
calcLandReport(project = "RESCUE", harmonizationPeriod = c(2015, 2050))
}
\arguments{
\item{project}{name of the project, currently only "RESCUE"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
land use data
}
\description{
Convert the downscaled land use data to the format required by the given project.
}
\author{
Pascal Sauer
}
