% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTargetLowRes.R
\name{calcLandTargetLowRes}
\alias{calcLandTargetLowRes}
\title{calcLandTargetLowRes}
\usage{
calcLandTargetLowRes(input = "magpie", target = "luh2mod")
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, currently only "luh2mod"}
}
\value{
low resolution target land data
}
\description{
Aggregate target land data to the spatial resolution of the input data in
preparation for harmonization.
}
\author{
Pascal Sauer
}
