% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandTransitions.R
\name{calcLandTransitions}
\alias{calcLandTransitions}
\title{calcLandTransitions}
\usage{
calcLandTransitions(project = "RESCUE", gross = TRUE)
}
\arguments{
\item{project}{name of the project, currently only "RESCUE"}

\item{gross}{either boolean or a magpie object containing bidirectional
transition shares relative to the area of the involved land pools (transition
divided by the area of the land pool in the "from" sub dimension). If set to
FALSE only net transitions will be returned. If set to TRUE an internal
gross transition estimate based on average gross transitions in LUH2 in the
period from 1995 to 2015 will be used.}
}
\value{
land use transition data
}
\description{
Ex-Post estimation of land use transitions based on land use state
information
}
\author{
Jan Philipp Dietrich
}
