% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHarmonized.R
\name{calcNonlandHarmonized}
\alias{calcNonlandHarmonized}
\title{calcNonlandHarmonized}
\usage{
calcNonlandHarmonized(
  input = "magpie",
  target = "luh2",
  harmonizeYear = 2015,
  finalYear = 2050,
  method = "extrapolateFade"
)
}
\arguments{
\item{input}{name of the input dataset, currently only "magpie"}

\item{target}{name of the target dataset, currently only "luh2"}

\item{harmonizeYear}{start of the harmonization period, years before
this are taken from the target dataset}

\item{finalYear}{end of the harmonization period, years after this are
taken from the input dataset}

\item{method}{harmonization method, see \code{\link{toolGetHarmonizer}} for available methods}
}
\value{
harmonized nonland data
}
\description{
Harmonize nonland input data to target data using the specified method, checking
data for consistency before returning.
}
\author{
Pascal Führlich
}
