% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHarmonizedCategories.R
\name{calcNonlandHarmonizedCategories}
\alias{calcNonlandHarmonizedCategories}
\title{calcNonlandHarmonizedCategories}
\usage{
calcNonlandHarmonizedCategories(input = "magpie", target = "luh2mod")
}
\arguments{
\item{input}{name of the input dataset, currently only "magpie"}

\item{target}{name of the target dataset, currently only "luh2"}
}
\value{
nonland data with target categories
}
\description{
Harmonize categories by mapping nonland input data categories to the categories of the nonland target dataset.
See \code{\link{calcLandHarmonizedCategories}} for an explanation of the mapping procedure.
}
\author{
Pascal Sauer
}
