% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHighRes.R
\name{calcNonlandHighRes}
\alias{calcNonlandHighRes}
\title{calcNonlandHighRes}
\usage{
calcNonlandHighRes(input = "magpie", target = "luh2")
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, currently only "luh2"}
}
\value{
downscaled nonland data
}
\description{
This function performs the downscaling: It calculates a high resolution dataset
from the low resolution input dataset and the high resolution target dataset.
}
\author{
Pascal Sauer
}
