% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandHighRes.R
\name{calcNonlandHighRes}
\alias{calcNonlandHighRes}
\title{calcNonlandHighRes}
\usage{
calcNonlandHighRes(
  input = "magpie",
  target = "luh2mod",
  harmonizationPeriod = c(2015, 2050)
)
}
\arguments{
\item{input}{name of an input dataset, currently only "magpie"}

\item{target}{name of a target dataset, currently only "luh2mod"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
downscaled nonland data
}
\description{
This function performs the downscaling: It calculates a high resolution dataset
from the low resolution input dataset and the high resolution target dataset.
}
\author{
Pascal Sauer
}
