% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandReport.R
\name{calcNonlandReport}
\alias{calcNonlandReport}
\title{calcNonlandReport}
\usage{
calcNonlandReport(project = "RESCUE", harmonizationPeriod = c(2015, 2050))
}
\arguments{
\item{project}{name of the project, currently only "RESCUE"}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
nonland data
}
\description{
Convert the downscaled nonland data to the format required by the given project.
}
\author{
Pascal Sauer
}
