% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandReport.R
\name{calcNonlandReport}
\alias{calcNonlandReport}
\title{calcNonlandReport}
\usage{
calcNonlandReport(outputFormat = "ESM", harmonizationPeriod = c(2015, 2050))
}
\arguments{
\item{outputFormat}{format in which the outputs should be prepared. Currently,
only "ESM" for earth system model compatible input data is available.}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
nonland data
}
\description{
Convert the downscaled nonland data to the format required by the given project.
}
\author{
Pascal Sauer
}
