% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonlandTarget.R
\name{calcNonlandTarget}
\alias{calcNonlandTarget}
\title{calcNonlandTarget}
\usage{
calcNonlandTarget(
  target = "luh2mod",
  years = seq(1995, 2015, 5),
  timestepLength = 5
)
}
\arguments{
\item{target}{name of the target dataset, currently only "luh2" and "luh2mod" are supported}
}
\value{
nonland target data
}
\description{
Prepare the high resolution nonland target dataset for
harmonization and downscaling, checking data for consistency before returning.
}
\author{
Pascal Sauer
}
