% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRESCUE.R
\name{fullRESCUE}
\alias{fullRESCUE}
\title{fullRESCUE}
\usage{
fullRESCUE(
  rev = NULL,
  ...,
  scenario = "",
  years = 1995:2100,
  compression = 2,
  interpolate = FALSE
)
}
\arguments{
\item{rev}{revision number of the data. If not provided the current date will be used instead}

\item{...}{reserved for future use}

\item{scenario}{scenario name to be included in filenames}

\item{years}{Only years from this vector will be included in the output files, if interpolate is TRUE,
all years given here will be included}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}

\item{interpolate}{boolean defining whether the data should be interpolated
to include all years given in the years argument}
}
\description{
Run the pipeline to generate harmonized and downscaled data to report for the RESCUE project.
Write .nc files, print full report on consistency checks and write it to report.log.
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
