% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdownscale-package.R
\docType{package}
\name{mrdownscale-package}
\alias{mrdownscale}
\alias{mrdownscale-package}
\title{mrdownscale: Downscale and harmonize MAgPIE output using LUH2 as reference data}
\description{
Downscale and harmonize MAgPIE output using LUH2 as reference data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrdownscale}
  \item \doi{10.5281/zenodo.11244475}
  \item Report bugs at \url{https://github.com/pik-piam/mrdownscale/issues}
}

}
\author{
\strong{Maintainer}: Pascal Sauer \email{pascal.sauer@pik-potsdam.de} (\href{https://orcid.org/0000-0002-6856-8239}{ORCID}) (Potsdam Institute for Climate Impact Research)

Authors:
\itemize{
  \item Jan Philipp Dietrich \email{dietrich@pik-potsdam.de} (\href{https://orcid.org/0000-0002-4309-6431}{ORCID}) (Potsdam Institute for Climate Impact Research)
}

}
