% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdownscale-package.R
\docType{package}
\name{mrdownscale-package}
\alias{mrdownscale}
\alias{mrdownscale-package}
\title{mrdownscale: Downscale and harmonize MAgPIE output using LUH2 as reference data}
\description{
Downscale and harmonize MAgPIE output using LUH2 as reference data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrdownscale}
}

}
\author{
\strong{Maintainer}: Pascal Führlich \email{pascal.fuehrlich@pik-potsdam.de}

}
