% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLowHigh.R
\name{plotLowHigh}
\alias{plotLowHigh}
\title{plotLowHigh}
\usage{
plotLowHigh(
  variable,
  year,
  range = c(0, 1),
  xlim = c(-180, 180),
  ylim = c(-90, 90)
)
}
\arguments{
\item{variable}{character, a regex that should match at least one variable
name in LandHighRes and LandHarmonized, if multiple are matched they are
summed up}

\item{year}{integer, year to plot}

\item{range}{range for plot legend}

\item{xlim}{min and max x coordinate to plot}

\item{ylim}{min and max y coordinate to plot}
}
\description{
Plot a land variable in low (after harmonization) and high resolution (after
downscaling) for a given year, unit is area share.
}
\examples{
\dontrun{
plotLowHigh("c3ann", 2040, range = c(0, 0.9),
            ylim = c(60, 30), xlim = c(-10, 60))
}

}
\author{
Pascal Sauer
}
