% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLUH2v2h.R
\name{readLUH2v2h}
\alias{readLUH2v2h}
\title{readLUH2v2h}
\usage{
readLUH2v2h(subtype)
}
\arguments{
\item{subtype}{one of states, management, transitions, cellArea}
}
\description{
Read LUH2v2h data. For the states subtype, the secma and secmb categories are removed.
For the management subtype, only the categories crpbf, rndwd, fulwd, fertl and irrig are read.
For the transitions subtype, only the wood harvest categories bioh and harv are read. To
match magpie semantics years are shifted by 1 when reading transitions.
}
