% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMagpie.R
\name{readMagpie}
\alias{readMagpie}
\title{readMagpie}
\usage{
readMagpie(subtype = "land")
}
\arguments{
\item{subtype}{type of data to be read in. Available options are land,
crop, woodHarvestWeight, woodHarvestArea and fertilizer}
}
\description{
Read function for data coming from the MAgPIE model.
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
