% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolDownscaleMagpieClassic.R
\name{toolDownscaleMagpieClassic}
\alias{toolDownscaleMagpieClassic}
\title{toolDownscaleMagpieClassic}
\usage{
toolDownscaleMagpieClassic(x, xTarget)
}
\arguments{
\item{x}{magpie dataset containing land to be downscaled}

\item{xTarget}{target land use dataset as SpatRaster for initialization year}
}
\value{
downscaled land use dataset
}
\description{
classic MAgPIE downscaling method using \code{luscale::interpolate2} as
downscaling function.
}
\author{
Jan Philipp Dietrich
}
