% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeExtrapolateFade.R
\name{toolHarmonizeExtrapolateFade}
\alias{toolHarmonizeExtrapolateFade}
\title{toolHarmonizeExtrapolateFade}
\usage{
toolHarmonizeExtrapolateFade(
  input,
  target,
  harmonizeYear,
  finalYear,
  constantSum
)
}
\arguments{
\item{input}{input data}

\item{target}{target data}

\item{harmonizeYear}{year in which the transition from target to input
data begins}

\item{finalYear}{year in which the transition shall be completed}

\item{constantSum}{boolean indicating whether the total sum over all layers
is suppossed to stay contstant (e.g. sum over all land types) or not.}
}
\description{
Tool function for creating a harmonized data set with a smooth transition
from historic target data to simulated input data.
}
\details{
A smooth transition is achieved in 2 steps: 1) extrapolating the target data
into the future so that it covers the whole transition period and 2) fade
over from the extrapolated target data to the input data.

Extrapolation of target data is achieved by taking the historically observed
growth rates of the different categories, dampen them (to avoid overestimation
of trends), apply them into the future and finally, if \code{constantSum} is
set to TRUE, normalize the result to the total sum of each spatial entity.

Please note that it is crucial to set the switch \code{constantSum} correctly
for the specific application as otherwise the results will be inconsistent.
}
\author{
Jan Philipp Dietrich
}
