% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeExtrapolateFade.R
\name{toolHarmonizeExtrapolateFade}
\alias{toolHarmonizeExtrapolateFade}
\title{toolHarmonizeExtrapolateFade}
\usage{
toolHarmonizeExtrapolateFade(input, target, harmonizeYear, finalYear)
}
\arguments{
\item{input}{name of the land input source to be used (default "magpie")}

\item{target}{name of the land target source to be used (default "luh2")}

\item{harmonizeYear}{year in which the transition from target to input
data begins}

\item{finalYear}{year in which the transition shall be completed}
}
\description{
Tool function for creating a harmonized land data set with a smooth tranistion
from historic target data to simulated input data.
}
\details{
A smooth transition is achieved in 2 steps: 1) extrapolating the target data
into the future so that it covers the whole transition period and 2) fade
over from the extrapolated target data to the input data.

Extrapolation of target data is achieved by taking the historically observed
growth rates of the different land categories, dampen them (to avoid overestimation
of trends), apply them into the future and finally normalize the result
to the total area of each spatial entity.
}
\author{
Jan Philipp Dietrich
}
