% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeFade.R
\name{toolHarmonizeFade}
\alias{toolHarmonizeFade}
\title{toolHarmonizeFade}
\usage{
toolHarmonizeFade(xInput, xTarget, harmonizationPeriod)
}
\arguments{
\item{xInput}{input data as magpie object}

\item{xTarget}{target data as magpie object}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
harmonized data set as magpie object with data from input for years
before the harmonization period, data from target for years after the
harmonization period and a smooth transition in between.
}
\description{
Tool function for creating a harmonized data set with a smooth s-shaped
transition from historic target data to simulated input data.
}
\author{
Jan Philipp Dietrich, Pascal Sauer
}
