% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeFade.R
\name{toolHarmonizeFade}
\alias{toolHarmonizeFade}
\title{toolHarmonizeFade}
\usage{
toolHarmonizeFade(input, target, harmonizationPeriod)
}
\arguments{
\item{input}{magpie object to fade to, usually model projections}

\item{target}{magpie object to fade from, usually historical data, this argument is called "target"
for consistency with argument names of other functions, it is somewhat misleading here,
because data after the harmonizationEnd will not be taken from target, but from input}

\item{harmonizationPeriod}{Two integer values, before the first given
year the target dataset is used, after the second given year the input
dataset is used, in between harmonize between the two datasets}
}
\value{
magpie object with harmonized data
}
\description{
Harmonize two datasets by fading from one to the other, using the following formula:
.inputFactor(year) * input[, year, ] + (1 - .inputFactor(year)) * target[, year, ]
where .inputFactor is (year - harmonizationStart) / (harmonizationEnd - harmonizationStart).
Requires both datasets to have all years including and between harmonizationStart and harmonizationEnd,
so it is not suitable for harmonization periods in the future if historical data is used.
}
\author{
Pascal Sauer
}
