% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonizeOffset.R
\name{toolHarmonizeOffset}
\alias{toolHarmonizeOffset}
\title{toolHarmonizeOffset}
\usage{
toolHarmonizeOffset(input, target, harmonizeYear, finalYear)
}
\arguments{
\item{input}{magpie object to fade to, usually model projections}

\item{target}{magpie object to fade from, usually historical data, this argument is called "target"
for consistency with argument names of other functions, it is somewhat misleading here,
because data after the finalYear will not be taken from target, but from input}

\item{harmonizeYear}{year to start fading, data at and before this year will be taken from target}

\item{finalYear}{year to stop fading, data at and after this year will be taken from input}
}
\value{
magpie object with harmonized data
}
\description{
Harmonize datasets using the offset method as implemented in mip::harmonize,
originally implemented in Python as part of aneris.
}
\author{
Pascal Führlich
}
