% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolPrimFix.R
\name{toolPrimFix}
\alias{toolPrimFix}
\title{toolPrimFix}
\usage{
toolPrimFix(x, noteThreshold = 10^-10, warnThreshold = 10^-5)
}
\arguments{
\item{x}{a magpie object with categories c("primf", "primn", "secdf", "secdn")}

\item{noteThreshold}{prim expansion greater than this will trigger a note
that primf/primn expansion was replaced by secdf/secdn expansion}

\item{warnThreshold}{prim expansion greater than this will trigger a warning
that primf/primn are expanding considerably}
}
\value{
a magpie object with primf/primn expansion replaced by secdf/secdn expansion
}
\description{
Replace primf and primn expansion with secdf and secdn expansion.
}
\author{
Pascal Sauer
}
