% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWriteNC.R
\name{toolWriteNC}
\alias{toolWriteNC}
\title{toolWriteNC}
\usage{
toolWriteNC(
  x,
  variables,
  fileName,
  now = Sys.time(),
  compression = 2,
  interpolationType = NULL,
  years = 1995:2100
)
}
\arguments{
\item{x}{magpie object to be written to file}

\item{variables}{layers that should be selected from \code{x}}

\item{fileName}{file name that should be used for the resulting nc file}

\item{now}{time that should be used as time stamp in metadata}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}

\item{interpolationType}{Either NULL (no interpolation), "linear" (linear interpolation), or "constant"
(last available year is being repeated)}

\item{years}{years to be returned. Will only be considered if a interpolation type is selected}
}
\description{
Tool function to write LUH2-style NetCDF files
}
\author{
Pascal Sauer, Jan Philipp Dietrich
}
