% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolWriteTransitions.R
\name{toolWriteTransitions}
\alias{toolWriteTransitions}
\title{toolWriteTransitions}
\usage{
toolWriteTransitions(trans, fileSuffix, now = Sys.time(), compression = 2)
}
\arguments{
\item{trans}{land transition data the information should be extracted from}

\item{fileSuffix}{second half of the resulting file name for the output NetCDF file}

\item{now}{time that should be used as time stamp in metadata}

\item{compression}{compression level of the resulting .nc files, possible values are integers from 1-9,
1 = fastest, 9 = best compression}
}
\description{
Tool function to write LUH2-style land transition data in NetCDF format
}
\author{
Jan Philipp Dietrich
}
